refactor: Presentation 계층에 config 폴더 생성
